requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
jupyter>=1.0.0
ipykernel>=6.0.0

# PDF processing
PyMuPDF>=1.18.0
pdfplumber>=0.7.0

# API-based extraction
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=2.0.0

# Small LM extraction
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.96

# Advanced NLP extraction
pytorch-crf>=0.7.2
spacy>=3.4.0
nltk>=3.7
rake-nltk>=1.0.6
textstat>=0.7.3
scikit-learn>=1.0.0

# Utilities
tqdm>=4.62.0
loguru>=0.6.0
jsonschema>=4.0.0
huggingface-hub>=0.16.0